+++ index.html
@@ -7,8 +7,8 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
+  <script type="module" src="js/main.js"></script>
   <script src="js/functions.js" async></script>
-  <script src="js/main.js" defer></script>
 </head>
 
 <body>
